+++ WalletWasabi.Tests/UnitTests/Clients/WasabiClientTests.cs
@@ -7,7 +7,6 @@ using System.Net;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
-using WalletWasabi.TorSocks5;
 using WalletWasabi.WebClients.Wasabi;
 using Xunit;
 
@@ -90,36 +89,6 @@ namespace WalletWasabi.Tests.UnitTests.Filters
 			tx.Outputs.Add(Money.Coins(2), Script.Empty);
 			tx.Outputs.Add(Money.Coins(1), Script.Empty);
 			return tx;
-		}
-	}
-
-	class MockTorHttpClient : ITorHttpClient
-	{
-		public Uri DestinationUri => new Uri("DestinationUri");
-
-		public Func<Uri> DestinationUriAction => () => DestinationUri;
-
-		public EndPoint TorSocks5EndPoint => IPEndPoint.Parse("127.0.0.1:9050");
-
-		public bool IsTorUsed => true;
-
-		public Func<HttpMethod, string, string[], Task<HttpResponseMessage>> OnSendAsync_Method { get; set; }
-
-		public Task<HttpResponseMessage> SendAsync(HttpMethod method, string relativeUri, HttpContent content = null, CancellationToken cancel = default)
-		{
-			var sepPos = relativeUri.IndexOf('?');
-			var action = relativeUri[..sepPos];
-			var parameters = relativeUri[(sepPos+1)..].Split('&', StringSplitOptions.RemoveEmptyEntries);
-			return OnSendAsync_Method(method, action, parameters);
-		}
-
-		public Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancel = default)
-		{
-			throw new NotImplementedException();
-		}
-
-		public void Dispose()
-		{
 		}
 	}
 }
